athena: CCTV credentials rework and add live image page
Don't hardcode credentials in source URLs. Instead use placeholders for
username and password.

A custom PreExceStart phase gets the credentials from systemd, and runs
the configuration template through envsubst to output the actual
configuration with the username and password. This is much easier than
the initial attempt of using the MediaMTX API, which had issues with
idempotency.

Create simple webpage using Caddy to show live snapshot and camera
settings. This uses the same credentials file as MediaMTX, but encoded
in Base64 for basic authentication.